# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an Api for Yolla App",
        "title": "Yolla Api",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.1.0"
    },
    "paths": {
        "/v1/users": {
            "get": {
                "description": "Gets all users from db",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get all users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/users/register": {
            "post": {
                "description": "Creates a new user account with the provided information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User registration data",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.RegisterUserPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "user created succesfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.RegisterUserPayload": {
            "type": "object",
            "required": [
                "email",
                "firstname",
                "lastname",
                "password",
                "phone",
                "username"
            ],
            "properties": {
                "age": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 14
                },
                "email": {
                    "type": "string",
                    "maxLength": 55
                },
                "firstname": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 3
                },
                "lastname": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 3
                },
                "password": {
                    "type": "string",
                    "maxLength": 25,
                    "minLength": 8
                },
                "phone": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 9
                },
                "username": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 4
                }
            }
        }
    }
}